+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0-beta6] - 2024-06-16
+
 ## Fixed
 
 * Automatically remove Tell Campfire Story effects after daily preparations